requirements.txt
pyyaml>=6.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
pillow>=10.0.0
transformers>=4.49.0
accelerate>=0.34.0
peft>=0.13.0
bitsandbytes>=0.43.0
datasets>=2.21.0
safetensors>=0.4.5
